pyproject.toml
[project]
name = "scale-test"
version = "0.1.0"
description = "AI Observability Scale Test Framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opentelemetry-api==1.25.0",
    "opentelemetry-sdk==1.25.0",
    "opentelemetry-exporter-otlp-proto-http==1.25.0",
    "aiohttp==3.9.5",
    "python-dotenv==1.0.1",
    "numpy==1.26.4",
    "pytest==8.2.0",
    "pytest-asyncio==0.23.7",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.0",
    "pytest-asyncio==0.23.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
